Bound type information look up to an action as well.
Alt - Equals will also bring up the type info, next to the
documentation provider underneath Ctrl - Q or hover.
